+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added `String.after()`
 - Added `String.afterLast()`
 - Added `String.before()`
+- Added `String.beforeLast()`
 
 ## 0.1.0